fix: moment format
