+++ src/business/PostsBusiness.ts
@@ -10,19 +10,6 @@ export class PostsBusiness {
     const postDatabase = new PostsDatabase();
     const response = await postDatabase.getFeed(authenticationData.id, type);
 
-    const posts: PostAndUserNameOutputDTO[] = response.map((item) => {
-      const post: PostAndUserNameOutputDTO = {
-        postId: item.post_id,
-        urlPhoto: item.utl_photo,
-        description: item.description,
-        creationDate: moment(item.creation_date).format("DD/MM/YYYY HH:mm"),
-        type: item.type,
-        userId: item.user_creator_id,
-        userName: item.user_creator_name,
-      };
-      return post;
-    });
-
-    return posts;
+    return response;
   }
 }
